requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
Mako==1.3.6
MarkupSafe==3.0.2
openai==1.54.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.0
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
